Fixed the errors in procedure line by forcing x to be an integer

+++ Activity 3.1.py
@@ -8,7 +8,7 @@ right_justify("Python is amazing!")
 
 
 def line(x):
-    x = ((x-3)/2)
+    x = int((x-3)/2)
     y = "+"+("-"*x)+"+"+("-"*x)+"+"
     print(y)
